version 1.5.0

+++ compiler/funcs.go
@@ -26,7 +26,7 @@ func newFunc(cmpl *compiler, name string) int {
 	funcObj.Block.Object = funcObj
 	cmpl.owners = append(cmpl.owners, &funcObj.Block) //[]core.ICmd{&funcObj.Block}
 	ind := cmpl.appendObj(funcObj)
-	funcObj.ObjID = uint16(ind)
+	funcObj.ObjID = int32(ind)
 	return ind
 }
